xe: rename repetitive file names
