+++ src/gpu/intel/ocl/bnorm/gen9_batch_normalization.cpp
@@ -18,7 +18,7 @@
 #include "common/experimental.hpp"
 #include "common/utils.hpp"
 #include "gpu/intel/compute/utils.hpp"
-#include "gpu/intel/ocl/bnorm/bnorm_utils.hpp"
+#include "gpu/intel/ocl/bnorm/utils.hpp"
 #include "gpu/intel/ocl/utils.hpp"
 
 using namespace dnnl::impl::memory_tracking::names;
