+++ src/gpu/intel/ocl/bnorm/lookup_table.cpp
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *******************************************************************************/
-#include "gpu/intel/ocl/bnorm/bnorm_lookup_table.hpp"
+#include "gpu/intel/ocl/bnorm/lookup_table.hpp"
 #include "gpu/intel/compute/compute_engine.hpp"
 
 #include <string>
