+++ src/gpu/intel/ocl/bnorm/model.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2023-2024 Intel Corporation
+* Copyright 2023-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -13,12 +13,12 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *******************************************************************************/
-#include "gpu/intel/ocl/bnorm/bnorm_model.hpp"
+#include "gpu/intel/ocl/bnorm/model.hpp"
 #include <cmath>
 #include "common/utils.hpp"
 #include "gpu/intel/compute/utils.hpp"
-#include "gpu/intel/ocl/bnorm/bnorm_utils.hpp"
 #include "gpu/intel/ocl/bnorm/nhwc_batch_normalization.hpp"
+#include "gpu/intel/ocl/bnorm/utils.hpp"
 
 namespace dnnl {
 namespace impl {
