+++ src/gpu/intel/ocl/bnorm/utils.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2023-2024 Intel Corporation
+* Copyright 2023-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -13,10 +13,10 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *******************************************************************************/
-#include "gpu/intel/ocl/bnorm/bnorm_utils.hpp"
 #include "common/utils.hpp"
 #include "gpu/intel/compute/utils.hpp"
-#include "gpu/intel/ocl/bnorm/bnorm_lookup_table.hpp"
+#include "gpu/intel/ocl/bnorm/lookup_table.hpp"
+#include "gpu/intel/ocl/bnorm/utils.hpp"
 
 namespace dnnl {
 namespace impl {
