+++ src/gpu/intel/ocl/reduction/atomic_reduction.cpp
@@ -28,7 +28,7 @@
 #include "gpu/intel/compute/utils.hpp"
 #include "gpu/intel/gpu_primitive_attr.hpp"
 #include "gpu/intel/ocl/reduction/atomic_reduction.hpp"
-#include "gpu/intel/ocl/reduction/reduction_utils.hpp"
+#include "gpu/intel/ocl/reduction/utils.hpp"
 #include "gpu/intel/ocl/utils.hpp"
 
 namespace dnnl {
