+++ src/gpu/intel/ocl/reduction/combined_reduction.hpp
@@ -20,7 +20,7 @@
 #include "common/primitive.hpp"
 #include "gpu/gpu_reduction_pd.hpp"
 #include "gpu/intel/gpu_primitive.hpp"
-#include "gpu/intel/ocl/reduction/reduction_utils.hpp"
+#include "gpu/intel/ocl/reduction/utils.hpp"
 #include "gpu/intel/primitive_conf.hpp"
 
 namespace dnnl {
