+++ src/gpu/intel/ocl/reduction/utils.hpp
