+++ src/gpu/intel/ocl/rnn/common.h
