+++ src/gpu/intel/ocl/rnn/grid.hpp
@@ -26,7 +26,7 @@
 #include "gpu/gpu_resource.hpp"
 #include "gpu/gpu_rnn_pd.hpp"
 #include "gpu/intel/gpu_primitive.hpp"
-#include "gpu/intel/ocl/rnn/rnn_utils.hpp"
+#include "gpu/intel/ocl/rnn/utils.hpp"
 #include "gpu/intel/ocl/utils.hpp"
 #include "gpu/intel/primitive_conf.hpp"
