+++ src/gpu/intel/ocl/rnn/utils.hpp
