+++ src/gpu/intel/sycl/engine.hpp
@@ -35,7 +35,7 @@
 #include "gpu/intel/sycl/compat.hpp"
 #include "gpu/intel/sycl/utils.hpp"
 
-#include "gpu/intel/sycl/sycl_interop_gpu_kernel.hpp"
+#include "gpu/intel/sycl/interop_kernel.hpp"
 
 namespace dnnl {
 namespace impl {
@@ -81,7 +81,7 @@ public:
 
         for (size_t i = 0; i < kernel_names.size(); i++) {
             if (!sycl_kernels[i]) continue;
-            CHECK(sycl_interop_gpu_kernel_t::make(
+            CHECK(interop_kernel_t::make(
                     kernels[i], *sycl_kernels[i], program_src));
         }
         return status::success;
@@ -113,7 +113,7 @@ public:
         VCHECK_KERNEL(gpu::intel::sycl::compat::make_kernel(
                               sycl_kernel, kernel_name, this, binary),
                 VERBOSE_KERNEL_CREATION_FAIL, kernel_name);
-        CHECK(sycl_interop_gpu_kernel_t::make(kernel, *sycl_kernel, {}));
+        CHECK(interop_kernel_t::make(kernel, *sycl_kernel, {}));
         return status::success;
     }
